Implement watchlist; MAJOR DB CHANGE.
The User model is now in an AutoOneToOneField relation with the new
model named Profile. AutoOneToOneField comes from a library named
django-annoying; the default OneToOneField errors when you try to access
it when it hasn't been created. Auto- creates one at first access
attempt.

The logic to add/remove to a user's watchlist resides in
posts.views.AddToWatchlist, served at `/watch` which is POST'ed at when
the user clicks the star. The initial status of the stars are decided by
a {% if post in favourites %} template tag, and the context variable
`favourites` is assigned at cemiyet.views.PostList.get_context_data.

It may or may not be better to set those initial stars also by js, by
adding a GET case to posts.views.AddToWatchlist.