Create popularity field in Post model.
The Post model has a BigIntegerField named popularity. It is an
exponential decaying running average as described here:
https://stackoverflow.com/questions/11128086/simple-popularity-algorithm

In posts.models there are a few utility functions now.
update_popularity(post, response) updates the popularity of `post` using
the creation time of `response`. This is called by PostingFormView at
the time of form creation for every parent of the new post.
reset_popularities() runs through the entire Post table to
recalculate each popularity anew; this is intended to be run from the
shell manually in case of parameter change or breakage.